+++ src/app/(auth)/signup/components/NicknamePassword.tsx
@@ -18,8 +18,8 @@ import axiosInstance from '@/app/api/axiosInstance';
 import confetti from 'canvas-confetti';
 import { AxiosError } from 'axios';
 import { IoChevronBackSharp } from 'react-icons/io5';
-import { EyeSlashFilledIcon } from '../../login/components/EyeSlashFilledIcon';
-import { EyeFilledIcon } from '../../login/components/EyeFilledIcon';
+import { EyeSlashFilledIcon } from '../../../components/Input/EyeSlashFilledIcon';
+import { EyeFilledIcon } from '../../../components/Input/EyeFilledIcon';
 
 interface EmailProps {
   validEmail: string;
